// projects/hr-smart/src/app/webservice/generated/hr-smart-service/model/departementEnum.ts
/**
 * HR-Smart Logging-Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type DepartementEnum = 'development' | 'Quality' | 'Design' | 'Other';

export const DepartementEnum = {
    Development: 'development' as DepartementEnum,
    Quality: 'Quality' as DepartementEnum,
    Design: 'Design' as DepartementEnum,
    Other: 'Other' as DepartementEnum
};